Update project settings and add page transition feature
